+++ python/ray/serve/api.py
@@ -6,8 +6,6 @@ import logging
 import random
 import re
 import time
-import json
-import yaml
 from dataclasses import dataclass
 from functools import wraps
 from typing import (
@@ -19,7 +17,6 @@ from typing import (
     Type,
     Union,
     List,
-    TextIO,
     overload,
 )
 
@@ -64,11 +61,7 @@ from ray.serve.utils import (
 from ray.util.annotations import PublicAPI
 import ray
 from ray import cloudpickle
-from ray.dashboard.modules.serve.schema import (
-    ServeApplicationSchema,
-    serve_application_to_schema,
-    schema_to_serve_application,
-)
+
 
 _INTERNAL_REPLICA_CONTEXT = None
 _global_client = None
@@ -1430,134 +1423,3 @@ def list_deployments() -> Dict[str, Deployment]:
         )
 
     return deployments
-
-
-def get_deployment_statuses() -> Dict[str, DeploymentStatusInfo]:
-    # Returns a dictionary of deployment names and their health statuses
-
-    return _get_global_client().get_deployment_statuses()
-
-
-def deploy_group(deployments: List[Deployment], _blocking: bool = True):
-    """
-    EXPERIMENTAL API
-
-    Takes in a list of deployment object, and deploys them atomically.
-
-    Args:
-        deployments(List[Deployment]): a list of deployments to deploy.
-        _blocking(bool): whether to wait for the deployments to finish
-            deploying or not.
-    """
-
-    if len(deployments) == 0:
-        return []
-
-    parameter_group = []
-
-    for deployment in deployments:
-        if not isinstance(deployment, Deployment):
-            raise TypeError(
-                f"deploy_group only accepts Deployments, but got unexpected "
-                f"type {type(deployment)}."
-            )
-
-        deployment_parameters = {
-            "name": deployment._name,
-            "func_or_class": deployment._func_or_class,
-            "init_args": deployment.init_args,
-            "init_kwargs": deployment.init_kwargs,
-            "ray_actor_options": deployment._ray_actor_options,
-            "config": deployment._config,
-            "version": deployment._version,
-            "prev_version": deployment._prev_version,
-            "route_prefix": deployment.route_prefix,
-            "url": deployment.url,
-        }
-
-        parameter_group.append(deployment_parameters)
-
-    _get_global_client().deploy_group(parameter_group, _blocking=_blocking)
-
-
-class Application:
-    def __init__(self, deployments: List[Deployment]):
-        self._deployments = {d.name: d for d in deployments}
-
-    def add_deployment(self, deployment: Deployment):
-        """Add deployment to the list. Validate name uniqueness."""
-
-        self._deployments[deployment.name] = deployment
-
-    def deploy(self, _blocking: bool = True):
-        """Async deploy (replace deploy_group())."""
-
-        return deploy_group(self._deployments, _blocking=_blocking)
-
-    def get_status(self) -> Dict[str, DeploymentStatus]:
-        """Get current status of all deployments."""
-
-        return get_deployment_statuses()
-
-    def run(self):
-        """Blocking run."""
-        pass
-
-    def __getitem__(self, key: str):
-        """Fetch deployment by name using dict syntax: app["name"]"""
-
-        if key in self._deployments:
-            return self._deployments[key]
-        else:
-            raise KeyError(
-                "Serve application does not contain a " f'"{key}" deployment.'
-            )
-
-    def __getattr__(self, name: str):
-        """Fetch deployment by name using attributes: app.name"""
-
-        if name in self._deployments:
-            return self._deployments[name]
-        else:
-            raise AttributeError(
-                "Serve application does not contain a " f'"{name}" deployment.'
-            )
-
-    def to_json(self, f: Optional[TextIO] = None) -> str:
-        """Write list of deployments to json str or file."""
-
-        json_str = serve_application_to_schema(self._deployments).json(indent=4)
-
-        if f:
-            f.write(json_str)
-        return json_str
-
-    @classmethod
-    def from_json(cls, str_or_file: Union[str, TextIO]) -> "Application":
-        """Load list of deployments from json str or file."""
-
-        if isinstance(str_or_file, str):
-            schema = ServeApplicationSchema.parse_raw(
-                str_or_file, content_type="application/json"
-            )
-        else:
-            schema = ServeApplicationSchema.parse_obj(json.load(str_or_file))
-
-        return Application(schema_to_serve_application(schema))
-
-    def to_yaml(self, f: Optional[TextIO]) -> Optional[str]:
-        """Write list of deployments to yaml str or file."""
-
-        json_str = serve_application_to_schema(self._deployments).json(indent=4)
-
-        if f:
-            yaml.safe_dump(json_str, f)
-        return yaml.safe_dump(json_str)
-
-    @classmethod
-    def from_yaml(cls, str_or_file: Union[str, TextIO]) -> "Application":
-        """Load list of deployments from yaml str or file."""
-
-        deployments_json = yaml.safe_load(str_or_file)
-        schema = ServeApplicationSchema.parse_obj(deployments_json)
-        return Application(schema_to_serve_application(schema))
